Finished firestore database access implementation

+++ tests/ValidPeople.UnitTests/Fakers/Entities/PersonFaker.cs
@@ -15,7 +15,7 @@ namespace ValidPeople.UnitTests.Fakers.Entities
                 ParentFaker.Get().Generate(2),
                 new Cpf(f.Person.Cpf(), f.Date.Future(), f.Date.Past()),
                 f.PickRandom(HobbyEnumeration.GetAll()),
-                f.Random.Decimal()));
+                f.Random.Double()));
         }
     }
 }
